+++ js/utils.js
@@ -14,22 +14,13 @@ const getRandomPositiveFloat = (min, max, digits = 1)=> {
   return result.toFixed(digits);
 };
 
-const declineOfNum = (num, expressions)=> {
-  let result;
-  let count = num % 100;
-  if (count >= 5 && count <= 20) {
-    result = expressions[2];
-  } else {
-    count = count % 10;
-    if (count === 1) {
-      result = expressions[0];
-    } else if (count >= 2 && count <= 4) {
-      result = expressions[1];
-    } else {
-      result = expressions[2];
-    }
+const declinationOfNum = (num, dict)=> {
+  if (num % 10 === 1 && num % 100 !== 11) {
+    return dict.single;
+  } else if (num % 10 >= 2 && num % 10 <= 4 && (num % 100 < 10 || num % 100 >= 20)) {
+    return dict.several;
   }
-  return result;
+  return dict.many;
 };
 
-export {getRandomPositiveFloat, getRandomPositiveInteger, declineOfNum};
+export {getRandomPositiveFloat, getRandomPositiveInteger, declinationOfNum};
